Completed map view.
Map tiles are cached onto a buffer bitmap. Only tiles which are visible and have changed since the last render are re-rendered to the buffer. Of course this means that maps are inherently limited to the players maximum bitmap size (4000x4000), and so a less optimized (yet also less limited) version of map view would be desirable as well.